Rename files to adhere with python standards
